+++ OSDSimulator/osd/src/osd_common.h
@@ -12,7 +12,7 @@ u32 osd_blend_pixel(u32 dst, u32 src, u8 alpha);
 
 int osd_line_style_check(u32 style, u32 x);
 
-u32 osd_color_add(u32 color, u8 r_delta, u8 g_delta, u8 b_delta);
+u32 osd_color_add(u32 color, int r_delta, int g_delta, int b_delta);
 
 void osd_merge_line(u32 *dst_buf, u32 *src_buf, u16 len, u32 x, u8 alpha);
